requirements.txt
Flask
Werkzeug
SQLAlchemy
flask_sqlalchemy
mysql-connector-python
